Add support for queuing entire spotify playlists

+++ plugins/MusicPlayer/player.js
@@ -1,6 +1,8 @@
 const YoutubeDL = require('youtube-dl');
 const Discord = require('discord.js');
 const MemoryStream = require('memorystream');
+const SpotifyWebApi = require('spotify-web-api-node');
+const AuthDetails = require("../../auth.js").getAuthDetails();
 
 let options = false;
 const MUSIC_CHANNEL_NAME = (options && options.musicChannelName) || 'music';
@@ -16,7 +18,8 @@ exports.commands = [
     "queue",
     "dequeue",
     "pause",
-    "resume"
+    "resume",
+    "playlist"
 ]
 
 function getResultTitle(result){
@@ -136,10 +139,18 @@ class Player {
             console.log("YoutubeDL Stream end");
         })
         setTimeout(()=>{
-            if(this.queue[0].response.channel){
-                const embed = generateResultEmbed('Now Playing',video_info,this.queue[0].queuer);
-                this.queue[0].response.channel.send('',embed);
-                this.queue[0].response.delete();
+            if(this.queue[0].response){
+                if(this.queue[0].response.channel){
+                    const embed = generateResultEmbed('Now Playing',video_info,this.queue[0].queuer);
+                    this.queue[0].response.channel.send('',embed);
+                    this.queue[0].response.delete();
+                }
+            } else {
+                let msg_channel = this.voiceChannel.guild.channels.cache.find((v)=>v.type == "text" && v.name === MUSIC_CHANNEL_NAME);
+                if(msg_channel){
+                    const embed = generateResultEmbed('Now Playing',video_info,this.queue[0].queuer);
+                    msg_channel.send('',embed);
+                }
             }
         const dispatcher = connection.play(buffer,{bitrate:'auto',volume:true});
         this.stream = stream;
@@ -340,5 +351,97 @@ exports.resume = {
         // Resume.
         msg.channel.send( wrap('Playback resumed.'));
         player.resume();
+    }
+}
+
+exports.playlist = {
+    description: "plays spotify playlists",
+    process: async function(client, msg, suffix) {
+        // Make sure the suffix exists.
+        if (!suffix) return msg.channel.send( wrap('No playlist specified!'));
+
+        // Make sure the user is in voice.
+        var arr = msg.guild.channels.cache.filter((v)=>v.type == "voice").filter((v)=>v.members.has(msg.author.id));
+        let responseChannel = msg.guild.channels.cache.find((v)=>v.type == "text" && v.name === MUSIC_CHANNEL_NAME) || msg.channel;
+        if (arr.length == 0) return msg.channel.send( wrap('You\'re not in a voice channel.'));
+
+        var show_playlist = true;
+
+        var playlist_id = null;
+        let uri_re = /spotify:playlist:(\w+)/i;
+        let uri = uri_re.exec(suffix);
+        if(uri) {
+            playlist_id = uri[1];
+        } else {
+            let link_re = /https:\/\/open.spotify.com\/playlist\/(\w+).*/i
+            let link = link_re.exec(suffix);
+            if(link) {
+                playlist_id = link[1];
+            }
+            //Don't show our playlist embed with an open spotify link since discord makes a nice embed.
+            show_playlist = false;
+        }
+        if(!playlist_id){
+            return msg.channel.send("This doesn't look like a spotify playlist to me...");
+        } else {
+            try {
+                var spotifyApi = new SpotifyWebApi({
+                    clientId: AuthDetails.spotify_client_id,
+                    clientSecret: AuthDetails.spotify_client_secret
+                });
+                console.log("requesting spotify access token");
+                let result = await spotifyApi.clientCredentialsGrant();
+                spotifyApi.setAccessToken(result.body['access_token']);
+                console.log("requesting playlist details for " + playlist_id);
+                let playlist = await spotifyApi.getPlaylist(playlist_id);
+                if(show_playlist){
+                    msg.channel.send("",{
+                        embed: {
+                            color: 0x1db954,
+                            author: {
+                                name: playlist.body.owner.display_name,
+                            },
+                            url: playlist.body.external_urls.spotify,
+                            image: {
+                                url: playlist.body.images[0].url,
+                            },
+                            title: playlist.body.name
+                        }
+                    });
+                }
+                // Now queue up the songs for playback
+                let player = getPlayer(msg.guild.id);
+                console.log("Starting search of youtube for tracks in this playlist...");
+                for(item of playlist.body.tracks.items) {
+                    let song_search = 'ytsearch1:' + item.track.name + ' ' + item.track.artists[0].name;
+                    try {
+                        let promise = new Promise((resolve, reject) => {
+                            YoutubeDL.getInfo(song_search, ['-i','--max-downloads', '1', '--no-playlist', '--no-check-certificate'], (err,info) =>{
+                                if(err){
+                                    reject(err);
+                                } else {
+                                    resolve(info);
+                                }
+                            });
+                        });
+
+                        let info = await promise;
+
+                        player.enqueue(client,msg,null,info);
+                    } catch (e) {
+                        console.log("failed trying to search for a song with " + song_search);
+                        conseole.error(e);
+                    }
+                }
+                console.log("All tracks in playlist" + playlist_id + " queued");
+            } catch(e) {
+                if(!AuthDetails.spotify_client_id || !AuthDetails.spotify_client_secret){
+                    console.log("Missing spotify api credentials. Did you add them to auth.json?");
+                    console.error(e);
+                }
+                return msg.channel.send("Couldn't read the playlist :(");
+            }
+        }
+        
     }
 }
